backend/requirements.txt
fastapi
uvicorn
langchain-core
langchain-ollama
python-dotenv
langserve
google-generativeai
langchain_community
sse_starlette
langchain-google-genai